This repository consists of Hybrid AES DES implementation in VHDL
AES and DES are two different crypto algorithms having different
features. This projects consists of integrating these algorithms to
develop a new structure. Here, read and write of text files is employed.
Thus, the text files listed should exist in the same folder as the
project is in. Implementation is carried in VHDL on Modelsim.